+++ source/others/FAST_INT64/f128_to_i64.cpp
@@ -78,8 +78,7 @@ f128_to_i64(float128_t a,
         sig64 |= UINT64_C(0x0001000000000000);
 
         if (shiftDist) {
-            /** @todo Warning   C4244   '=': conversion from 'int32_t' to 'uint8_t', possible loss of data */
-            sig128 = softfloat_shortShiftLeft128(sig64, sig0, -shiftDist);
+            sig128 = softfloat_shortShiftLeft128(sig64, sig0, static_cast<uint8_t>(-shiftDist));
             sig64 = sig128.v64;
             sig0 = sig128.v0;
         }
@@ -90,7 +89,7 @@ f128_to_i64(float128_t a,
             sig64 |= UINT64_C(0x0001000000000000);
         }
 
-        sigExtra = softfloat_shiftRightJam64Extra(sig64, sig0, shiftDist);
+        sigExtra = softfloat_shiftRightJam64Extra(sig64, sig0, static_cast<uint32_t>(shiftDist));
         sig64 = sigExtra.v;
         sig0 = sigExtra.extra;
     }
