+++ source/others/FAST_INT64/f32_to_ui64.cpp
@@ -66,7 +66,7 @@ f32_to_ui64(float32_t a,
     uint64_t extra = 0;
 
     if (shiftDist) {
-        uint64_extra const sig64Extra = softfloat_shiftRightJam64Extra(sig64, 0, shiftDist);
+        uint64_extra const sig64Extra = softfloat_shiftRightJam64Extra(sig64, 0, static_cast<uint32_t>(shiftDist));
         sig64 = sig64Extra.v;
         extra = sig64Extra.extra;
     }
